added binding extension and added BaseChatFragment. Now the structure is ChildChatFragment -> BaseChatFragment -> BaseFragment
