+++ app/src/main/java/com/space/chatApp/presentation/chat_activity/activity/ChatActivity.kt
@@ -5,9 +5,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.lifecycle.lifecycleScope
 import com.space.chat.databinding.ActivityMainBinding
-import com.space.chatApp.domain.model.UserType
 import com.space.chatApp.presentation.chat_activity.viewModel.ChatActivityViewModel
-import com.space.chatApp.presentation.chat_screen.ui.ChatFragment
+import com.space.chatApp.presentation.chat_screen.ui.ChatFragmentFirst
+import com.space.chatApp.presentation.chat_screen.ui.ChatFragmentSecond
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -26,15 +26,14 @@ class ChatActivity : AppCompatActivity() {
         if (savedInstanceState == null){
             fragmentSetup()
         }
-
         changeThemeModes()
 
     }
 
     private fun fragmentSetup() {
         supportFragmentManager.beginTransaction().apply {
-            replace(binding.topFragment.id, ChatFragment(), UserType.FIRST_USER_ID)
-            replace(binding.bottomFragment.id, ChatFragment(), UserType.SECOND_USER_ID)
+            binding.firstFragment?.let { replace(it.id, ChatFragmentFirst()) }
+            binding.secondFragment?.let { replace(it.id, ChatFragmentSecond()) }
         }.commit()
     }
