+++ app/src/main/res/layout-land/activity_main.xml
@@ -8,7 +8,7 @@
     tools:context="com.space.chatApp.presentation.chat_activity.activity.ChatActivity">
 
     <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/topFragment"
+        android:id="@+id/firstFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_weight="1"
@@ -28,7 +28,7 @@
         app:layout_constraintTop_toTopOf="parent"/>
 
     <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/bottomFragment"
+        android:id="@+id/secondFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_weight="1"
