+++ app/src/main/res/layout/activity_main.xml
@@ -17,7 +17,7 @@
         app:track="@drawable/ic_track" />
 
     <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/topFragment"
+        android:id="@+id/firstFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="1"  />
@@ -28,7 +28,7 @@
         app:dividerColor="@color/divider_primary" />
 
     <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/bottomFragment"
+        android:id="@+id/secondFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="1" />
